Adding test cases and code changes

+++ project0/main.py
@@ -8,16 +8,16 @@ def main(url):
     incident_data = fetchincidents(url)
         # Extract data
     incidents = extractincidents(incident_data)
-     # Create new database
+        # Create new database
     db = createdb()
-                            
-     #Insert data
-    populatedb(db, incidents)
-                                    
-    # Print incident counts
+        # Insert data
+    populatedb(db, incidents)                                    
+        # Print incident counts
     status(db)
 
 def fetchincidents(url):
+    if(url == ""):
+        raise Exception
     headers = {}
     headers['User-Agent'] = "Mozilla/5.0 (X11; Linux i686) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1312.27 Safari/537.17"                          
     data = urllib.request.urlopen(urllib.request.Request(url, headers=headers)).read()
@@ -54,13 +54,12 @@ def extractincidents(data):
             incident=page.split('\n')
             incident=incident[:-1]
             incidents.extend(incident)
+    print(type(incidents))
     return incidents    
 
 def createdb():
     db='normanpd.db'
-    #conn = None
     conn = sqlite3.connect('normanpd.db')
-    #conn.isolation_level = None
     cur = conn.cursor()
     cur.execute('''CREATE TABLE IF NOT EXISTS incidents (
                    incident_time TEXT,
@@ -76,6 +75,8 @@ def createdb():
 def  populatedb(db,incidents):
     incident_data=incidents
     db = 'normanpd.db'
+    if not (db == 'normanpd.db'and type(incidents == list) ):
+        raise Exception
     list_rows = list()
     z=0
     while( z<len(incident_data)):
@@ -83,20 +84,24 @@ def  populatedb(db,incidents):
          list_rows.append(list(incident_data[z:z+5]))
          z+=5
     con = sqlite3.connect('normanpd.db')
+    if( bool(con) == False):
+        raise Exception
     cur = con.cursor()
     #print(list_rows)
     for i in list_rows:
         cur.execute("INSERT into incidents VALUES (?,?,?,?,?)" , (i))
-    #sqlquery = 'INSERT into incidents VALUES (?,?,?,?,?)'
-    #cur.executemany(sqlquery,list_rows)
     con.commit()
     con.close()
 
 def status(db):
     db= 'normanpd.db'
     con = sqlite3.connect('normanpd.db')
+    if( bool(con) == False):
+        raise Exception
     cur = con.cursor()
     rows =  cur.execute("SELECT nature,COUNT(nature) FROM incidents GROUP BY nature ORDER BY nature")
+    if( bool(rows) == False):
+        raise Exception
     for row in rows:
         print(row[0],"|",row[1])
     cur.execute("DROP TABLE incidents")
